# docs/superpowers/specs/2026-04-21-deck-go-control-plane-api-openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Deck Control Plane API
  version: 0.1.0-stage2
  summary: Stage 2 target query/command HTTP contract for the control plane
servers:
  - url: /api/v1
paths:
...
  /alerts:
    get:
      summary: Fetch alert rules
      operationId: listAlertRules
      responses:
        "200":
          description: Alert rule list payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Create an alert rule
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Alert rule create payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /alerts/{ruleId}:
    patch:
      summary: Update an alert rule
      operationId: updateAlertRule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Alert rule update payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      summary: Delete an alert rule
      operationId: deleteAlertRule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert rule delete payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /docs/extract:
    post:
      summary: Extract docs from a session transcript
      operationId: extractDocs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionKey:
                  type: string
              required: [sessionKey]
      responses:
        "200":
          description: Docs extract payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /memory/search:
    get:
      summary: Search memory
      operationId: searchMemory
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Memory search payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /chat/session-events:
    post:
      summary: Subscribe or unsubscribe from chat session events
      operationId: updateChatSessionEventSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Session event subscription payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /runtimes/{runtimeId}:
    get:
      summary: Fetch one runtime summary
      operationId: getRuntime
      parameters:
        - $ref: "#/components/parameters/RuntimeId"
      responses:
        "200":
          description: Runtime summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  runtime:
                    $ref: "#/components/schemas/RuntimeSummary"
                  requestId:
                    type: string
                required: [runtime, requestId]
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    RuntimeId:
      name: runtimeId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Standard control-plane error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    RuntimeSummary:
      type: object
      properties:
        runtimeId:
          type: string
        managed:
          type: boolean
        configured:
          type: boolean
        status:
          type: string
        health:
          type: string
        capabilityVersion:
          type: string
          nullable: true
        methodCount:
          type: integer
          nullable: true
        eventCount:
          type: integer
          nullable: true
        gatewayUrl:
          type: string
          nullable: true
        lastError:
          type: string
          nullable: true
        autoStart:
          type: boolean
        occurredAt:
          type: string
          format: date-time
      required: [runtimeId, status, health, occurredAt]
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
          required: [code, message, details]
        requestId:
          type: string
      required: [error, requestId]